+++ packages/workspace-fusion/src/lib/integrations/common/components/TabNavigation.tsx
@@ -3,6 +3,7 @@ import { WorkspaceTabNames } from '../../../types';
 import { TabButtonDivider, TabButtonList, TabButton } from '../../../components/Header';
 import { useWorkspaceHeaderComponents } from '../../../context/WorkspaceHeaderComponents';
 import { Fragment, MutableRefObject, useRef } from 'react';
+import styled from 'styled-components';
 
 export const TabNavigation = () => {
 	const { icons, analyticsTabs, viewTabs } = useWorkspaceHeaderComponents();
@@ -10,21 +11,27 @@ export const TabNavigation = () => {
 
 	const setActiveTab = useSetActiveTab();
 
-	const leftIcons = icons.filter((s) => s.placement === 'left');
-	const rightIcons = icons.filter((s) => s.placement === 'right');
+	const leftIcons = icons.filter(({ placement }) => placement === 'left');
+	const rightIcons = icons.filter(({ placement }) => placement === 'right');
 
 	const activeTab = useActiveTab();
 	return (
 		<TabButtonList ref={pRef as MutableRefObject<HTMLDivElement>} id="button_list">
 			{!!leftIcons.length && (
 				<>
-					{leftIcons.map(({ Icon, name }) => (
+					{leftIcons.map(({ Icon, name, type }) => (
 						<>
 							{pRef.current && (
 								<Fragment key={name}>
-									<TabButton isActive={false}>
-										<Icon anchor={pRef.current} />
-									</TabButton>
+									{type === 'button' ? (
+										<TabButton isActive={false}>
+											<Icon anchor={pRef.current} />
+										</TabButton>
+									) : (
+										<StyledTextWrapper>
+											<Icon anchor={pRef.current} />
+										</StyledTextWrapper>
+									)}
 								</Fragment>
 							)}
 						</>
@@ -65,13 +72,19 @@ export const TabNavigation = () => {
 
 			{!!rightIcons.length && (
 				<>
-					{rightIcons.map(({ Icon, name }) => (
+					{rightIcons.map(({ Icon, name, type }) => (
 						<>
 							{pRef.current && (
 								<Fragment key={name}>
-									<TabButton isActive={false}>
-										<Icon anchor={pRef.current} />
-									</TabButton>
+									{type === 'button' ? (
+										<TabButton isActive={false}>
+											<Icon anchor={pRef.current} />
+										</TabButton>
+									) : (
+										<StyledTextWrapper>
+											<Icon anchor={pRef.current} />
+										</StyledTextWrapper>
+									)}
 								</Fragment>
 							)}
 						</>
@@ -81,3 +94,10 @@ export const TabNavigation = () => {
 		</TabButtonList>
 	);
 };
+
+const StyledTextWrapper = styled.div`
+	height: 100%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+`;