+++ packages/workspace-fusion/src/modules/garden/components/index.ts
@@ -1,0 +1,2 @@
+export * from './workspace-header';
+export * from './wrapper';